+++ src/ogx_client/types/__init__.py
@@ -13,19 +13,17 @@ from .shared import (
     ProviderInfo as ProviderInfo,
     SystemMessage as SystemMessage,
     SamplingParams as SamplingParams,
-    SafetyViolation as SafetyViolation,
     InterleavedContent as InterleavedContent,
     ListRoutesResponse as ListRoutesResponse,
     ListProvidersResponse as ListProvidersResponse,
     InterleavedContentItem as InterleavedContentItem,
 )
-from .shield import Shield as Shield
 from .vector_store import VectorStore as VectorStore
-from .create_response import CreateResponse as CreateResponse
 from .response_object import ResponseObject as ResponseObject
 from .file_list_params import FileListParams as FileListParams
 from .response_message import ResponseMessage as ResponseMessage
 from .batch_list_params import BatchListParams as BatchListParams
+from .model_list_params import ModelListParams as ModelListParams
 from .route_list_params import RouteListParams as RouteListParams
 from .compacted_response import CompactedResponse as CompactedResponse
 from .file_create_params import FileCreateParams as FileCreateParams
@@ -33,28 +31,25 @@ from .batch_create_params import BatchCreateParams as BatchCreateParams
 from .batch_list_response import BatchListResponse as BatchListResponse
 from .conversation_object import ConversationObject as ConversationObject
 from .list_files_response import ListFilesResponse as ListFilesResponse
+from .model_list_response import ModelListResponse as ModelListResponse
 from .route_list_response import RouteListResponse as RouteListResponse
-from .run_shield_response import RunShieldResponse as RunShieldResponse
 from .delete_file_response import DeleteFileResponse as DeleteFileResponse
 from .list_models_response import ListModelsResponse as ListModelsResponse
 from .prompt_create_params import PromptCreateParams as PromptCreateParams
 from .prompt_list_response import PromptListResponse as PromptListResponse
 from .prompt_update_params import PromptUpdateParams as PromptUpdateParams
 from .response_list_params import ResponseListParams as ResponseListParams
-from .shield_list_response import ShieldListResponse as ShieldListResponse
 from .batch_cancel_response import BatchCancelResponse as BatchCancelResponse
 from .batch_create_response import BatchCreateResponse as BatchCreateResponse
 from .chat_completion_chunk import ChatCompletionChunk as ChatCompletionChunk
 from .file_content_response import FileContentResponse as FileContentResponse
 from .list_prompts_response import ListPromptsResponse as ListPromptsResponse
-from .list_shields_response import ListShieldsResponse as ListShieldsResponse
 from .query_chunks_response import QueryChunksResponse as QueryChunksResponse
 from .prompt_retrieve_params import PromptRetrieveParams as PromptRetrieveParams
 from .provider_list_response import ProviderListResponse as ProviderListResponse
 from .response_create_params import ResponseCreateParams as ResponseCreateParams
 from .response_list_response import ResponseListResponse as ResponseListResponse
 from .response_object_stream import ResponseObjectStream as ResponseObjectStream
-from .shield_register_params import ShieldRegisterParams as ShieldRegisterParams
 from .vector_io_query_params import VectorIoQueryParams as VectorIoQueryParams
 from .batch_retrieve_response import BatchRetrieveResponse as BatchRetrieveResponse
 from .embedding_create_params import EmbeddingCreateParams as EmbeddingCreateParams
@@ -62,9 +57,7 @@ from .model_retrieve_response import ModelRetrieveResponse as ModelRetrieveRespo
 from .response_compact_params import ResponseCompactParams as ResponseCompactParams
 from .vector_io_insert_params import VectorIoInsertParams as VectorIoInsertParams
 from .completion_create_params import CompletionCreateParams as CompletionCreateParams
-from .moderation_create_params import ModerationCreateParams as ModerationCreateParams
 from .response_delete_response import ResponseDeleteResponse as ResponseDeleteResponse
-from .safety_run_shield_params import SafetyRunShieldParams as SafetyRunShieldParams
 from .vector_store_list_params import VectorStoreListParams as VectorStoreListParams
 from .completion_create_response import CompletionCreateResponse as CompletionCreateResponse
 from .conversation_create_params import ConversationCreateParams as ConversationCreateParams